requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
black==24.4.2
cachetools==5.3.3
camel-converter==3.1.2
celery==5.4.0
certifi==2024.6.2
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.4
cron-descriptor==1.4.5
cryptography==43.0.1
Deprecated==1.2.18
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-celery-beat==2.7.0
django-colorfield==0.11.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-editorjs-fields==0.2.7
django-environ==0.11.2
django-extensions==3.2.3
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
et_xmlfile==2.0.0
filelock==3.14.0
flake8==7.1.0
furl==2.1.3
identify==2.5.36
idna==3.7
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
lxml==5.3.0
mccabe==0.7.0
meilisearch==0.31.3
multidimensional_urlencode==0.0.4
mypy==1.10.0
mypy-extensions==1.0.0
netaddr==1.3.0
nodeenv==1.8.0
nose==1.3.7
numpy==2.1.3
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==24.0
pandas==2.2.3
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pook==2.1.1
pre-commit==3.7.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
pyTelegramBotAPI==4.24.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
transliterate==1.10.2
types-cachetools==5.5.0.20240820
types-requests==2.32.0.20240712
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.17.2
xmltodict==0.14.2
